Extract Fixture module into its own file
